chore(mobile): move metadata manipulation to asset service
