+++ packages/services/src/assets/stacks-asset.utils.ts
@@ -1,31 +1,11 @@
-import { hexToCV } from '@stacks/transactions';
-
 import {
   CryptoAssetCategories,
   CryptoAssetChains,
   CryptoAssetProtocols,
-  Sip9Asset,
-  Sip9Details,
   Sip10Asset,
-  SupportedSip9ContentType,
 } from '@leather.io/models';
 import { getTicker, isUndefined } from '@leather.io/utils';
 
-import { GammaNftMetadata } from '../infrastructure/api/gamma/gamma-api.client';
-import {
-  transformGammaAssetContent,
-  transformGammaAttributes,
-  transformGammaCollection,
-  transformGammaOwner,
-} from '../infrastructure/api/gamma/gamma-api.utils';
-import {
-  HiroMetadata,
-  HiroNftMetadataResponse,
-} from '../infrastructure/api/hiro/hiro-stacks-api.types';
-import {
-  mapHiroAttributes,
-  mapHiroCollection,
-} from '../infrastructure/api/hiro/hiro-stacks-api.utils';
 import { LeatherApiSip10Token } from '../infrastructure/api/leather/leather-api.client';
 
 export function isTransferableSip10Token(token: LeatherApiSip10Token) {
@@ -58,88 +38,6 @@ export function getContractPrincipalFromAddressAndName(
   contractName: string
 ) {
   return `${contractAddress}.${contractName}`;
-}
-
-export function getNonFungibleTokenId(hex: string): number {
-  const clarityValue = hexToCV(hex);
-  return clarityValue.type === 'uint' ? Number(clarityValue.value) : 0;
-}
-
-export function mapToSip9Details(
-  assetIdentifier: string,
-  gammaMetadata?: GammaNftMetadata | null,
-  hiroMetadata?: HiroMetadata | null
-): Sip9Details {
-  const assetName = getAssetNameFromIdentifier(assetIdentifier);
-
-  // Prioritize Gamma, fallback to Hiro
-  const name = gammaMetadata?.item.name || hiroMetadata?.name || assetName;
-
-  const description = gammaMetadata?.item.description || hiroMetadata?.description || '';
-
-  const cachedImage =
-    gammaMetadata?.item.asset_content?.content_url ||
-    hiroMetadata?.cached_image ||
-    hiroMetadata?.image ||
-    '';
-
-  const cachedImageThumbnail =
-    gammaMetadata?.item.asset_content?.content_url ||
-    (hiroMetadata as any)?.cached_thumbnail_image ||
-    hiroMetadata?.cached_image ||
-    '';
-
-  const contentType = (gammaMetadata?.item.asset_content?.content_type ||
-    '') as SupportedSip9ContentType;
-
-  const collection =
-    transformGammaCollection(gammaMetadata?.item.collection) ||
-    mapHiroCollection(hiroMetadata?.properties?.collection);
-
-  const owner = transformGammaOwner(gammaMetadata?.item.owner);
-
-  const attributes =
-    transformGammaAttributes(gammaMetadata?.attribute_groups) ||
-    mapHiroAttributes(hiroMetadata?.attributes);
-
-  return {
-    id: gammaMetadata?.item.id,
-    name,
-    description,
-    assetContent: transformGammaAssetContent(gammaMetadata?.item.asset_content),
-    cachedImage,
-    cachedImageThumbnail,
-    contentType,
-    locationUrl: gammaMetadata?.item.location_url,
-    collection,
-    owner,
-    attributes,
-    rarityRank: gammaMetadata?.item.rarity_rank,
-  };
-}
-
-export function createSip9Asset(
-  assetIdentifier: string,
-  tokenId: number,
-  metadata: HiroNftMetadataResponse | null,
-  gammaMetadata?: GammaNftMetadata | null
-): Sip9Asset {
-  const details = mapToSip9Details(assetIdentifier, gammaMetadata, metadata?.metadata);
-
-  return {
-    chain: CryptoAssetChains.stacks,
-    category: CryptoAssetCategories.nft,
-    protocol: CryptoAssetProtocols.sip9,
-    assetId: assetIdentifier,
-    contractId: getContractPrincipalFromAssetIdentifier(assetIdentifier),
-    tokenId,
-    name: details.name,
-    description: details.description,
-    cachedImage: details.cachedImage,
-    cachedImageThumbnail: details.cachedImageThumbnail,
-    contentType: details.contentType,
-    details,
-  };
 }
 
 export function createSip10Asset(sip10Token: LeatherApiSip10Token): Sip10Asset {
